requirements.txt
aioredis==1.3.1
aliyun-python-sdk-core-v3==2.9.0
aliyun-python-sdk-dysmsapi==1.0.0
amqp==2.2.2
appdirs==1.4.3
asgiref==3.2.7
async-timeout==2.0.1
atomicwrites==1.1.5
attrs==19.3.0
autobahn==20.4.1
Automat==20.2.0
autopep8==1.5
billiard==3.5.0.3
celery==4.1.0
certifi==2017.11.5
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
Click==7.0
confusable-homoglyphs==2.0.2
constantly==15.1.0
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
defusedxml==0.5.0
distlib==0.3.0
Django==3.0
django-3-jet==1.0.8
django-admin-bootstrapped==2.5.7
django-allauth==0.34.0
django-appconf==1.0.2
django-braces==1.13.0
django-ckeditor==5.9.0
django-cors-headers==3.2.1
django-crispy-forms==1.7.2
django-filter==1.1.0
django-formtools==2.1
django-grappelli==2.11.1
django-guardian==1.4.9
django-js-asset==1.2.2
django-messages==0.5.3
django-model-utils==3.0.0
django-modeladmin-reorder==0.3.1
django-notifs==2.6.0
django-oauth-toolkit==1.0.0
django-private-chat==0.1.9
django-progressbarupload==0.1.7
django-redis==4.9.0
django-registration==2.3
django-rest-auth==0.9.3
django-rest-swagger==2.0.7
django-silk==4.0.1
django-templated-mail==1.0.0
django-users2==0.2.1
django-xadmin==0.5.0
djangorestframework==3.11.0
djangorestframework-bulk==0.2.1
djangorestframework-jwt==1.11.0
djoser==1.1.5
drf-extensions==0.3.1
drf-nested-routers==0.90.0
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.0.2
gevent==1.4.0
gprof2dot==2019.11.30
greenlet==0.4.15
gunicorn==20.0.4
h11==0.9.0
hiredis==1.0.1
httptools==0.1.1
hyperlink==19.0.0
idna==2.6
importlib-metadata==1.5.0
importlib-resources==1.3.1
incremental==17.5.0
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
kombu==4.1.0
MarkupSafe==1.0
msgpack==0.6.2
mysqlclient==1.3.13
oauthlib==2.0.6
openapi-codec==1.3.2
pika==0.11.2
Pillow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pycryptodome==3.6.4
pycryptodomex==3.9.4
Pygments==2.6.1
PyHamcrest==2.0.2
PyJWT==1.6.1
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-alipay-sdk==2.0.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2017.3
redis==2.10.6
requests==2.18.4
requests-oauthlib==0.8.0
service-identity==18.1.0
simplejson==3.13.2
six==1.11.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
uritemplate==3.0.0
urllib3==1.22
uvicorn==0.11.3
uvloop==0.14.0
uWSGI==2.0.18
vine==1.1.4
virtualenv==20.0.8
websockets==8.1
Werkzeug==0.14.1
zipp==3.1.0
zope.interface==5.1.0
